Use long type for position in addPlayback:indexPath:position
This is the natural type used.

+++ XBMC Remote/DetailViewController.m
@@ -1511,7 +1511,7 @@
                     [self showActionSheet:indexPath sheetActions:sheetActions item:item rectOriginX:rectOriginX rectOriginY:rectOriginY];
                 }
                 else {
-                    [self addPlayback:item indexPath:indexPath position:(int)indexPath.row shuffle:NO];
+                    [self addPlayback:item indexPath:indexPath position:indexPath.row shuffle:NO];
                 }
                 return;
             }
@@ -1651,7 +1651,7 @@
                 [self showActionSheet:indexPath sheetActions:sheetActions item:item rectOriginX:rectOriginX rectOriginY:rectOriginY];
             }
             else {
-                [self addPlayback:item indexPath:indexPath position:(int)indexPath.row shuffle:NO];
+                [self addPlayback:item indexPath:indexPath position:indexPath.row shuffle:NO];
             }
         }
     }
@@ -3225,7 +3225,7 @@
         [self showActionSheetOptions:title options:sheetActions recording:isRecording origin:sheetOrigin fromcontroller:showFromCtrl fromview:self.view];
     }
     else if (indexPath != nil) { // No actions found, revert back to standard play action
-        [self addPlayback:item indexPath:indexPath position:(int)indexPath.row shuffle:NO];
+        [self addPlayback:item indexPath:indexPath position:indexPath.row shuffle:NO];
         forceMusicAlbumMode = NO;
     }
 }
@@ -3476,7 +3476,7 @@
             [self startSlideshow:item indexPath:selectedIndexPath];
         }
         else {
-            int playFromPosition = albumView ? (int)selectedIndexPath.row : 0;
+            long playFromPosition = albumView ? selectedIndexPath.row : 0;
             [self addPlayback:item indexPath:selectedIndexPath position:playFromPosition shuffle:NO];
         }
     }
@@ -4196,7 +4196,7 @@
     }];
 }
 
-- (void)addPlayback:(NSDictionary*)item indexPath:(NSIndexPath*)indexPath position:(int)pos shuffle:(BOOL)shuffled {
+- (void)addPlayback:(NSDictionary*)item indexPath:(NSIndexPath*)indexPath position:(long)pos shuffle:(BOOL)shuffled {
     mainMenu *menuItem = [self getMainMenu:item];
     NSDictionary *mainFields = menuItem.mainFields[choosedTab];
     if (forceMusicAlbumMode) {
